requirements.txt
certifi==2023.5.7
charset-normalizer==2.0.12
click==8.0.4
cycler==0.11.0
Flask==2.0.3
Flask-Cors==3.0.10
idna==3.4
importlib-metadata==4.8.3
itsdangerous==2.0.1
Jinja2==3.0.3
joblib==1.1.1
kiwisolver==1.3.1
MarkupSafe==2.0.1
matplotlib==3.3.4
numpy==1.19.5
pandas==1.1.5
Pillow==8.4.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2023.3
requests==2.27.1
scikit-learn==0.24.2
scipy==1.5.4
seaborn==0.11.2
six==1.16.0
threadpoolctl==3.1.0
typing==3.7.4.3
typing-extensions==4.1.1
urllib3==1.26.16
Werkzeug==2.0.3
zipp==3.6.0
